[FEAT] 게시글 api, S3 연동 및 관련 api 추가

+++ src/main/java/kdt/web_ide/post/dto/PostRequestDto.java
@@ -2,11 +2,10 @@ package kdt.web_ide.post.dto;
 
 import kdt.web_ide.post.entity.Language;
 import lombok.Data;
-import org.springframework.web.multipart.MultipartFile;
 
 @Data
 public class PostRequestDto {
-    private Integer boardId;
-    private String name;
-    private Language language;
+    private Long boardId; // 게시판 ID
+    private String name; // 게시글 이름
+    private Language language; // 언어 종류
 }
